Move config management to viper
We had our own implementation to read and parse YAML config files. But
in the future, we'll be passing secrets from ENV (that will override
values in the config). Handling that use-case will involve too much
boilerplate code. So, instead of writing and maintaining a crappy
solution for config management, we can use a battle-tested library
that is available for the same purpose.