Remove the use of `b` in global rate-limiter
As far as the global rate-limiter is concerned, all we want is to cap
requests at, say, x reqs per second. The rate-limiter internally uses a
token-bucket with the params `r` (rate at which bucket fills) and `b`
(max capacity of the bucket). For the global rate-limiter, to achieve
the above mentioned, both `r` and `b` must be equal. Having two params
in the config for the global rate-limiter can cause confusion
and improper configuration. So, this commit unifies it under a single
entry.

However, the interface for rate-limiter still supports passing two
separate `r` and `b`. So, if needed, it can be used so.